+++ src/RecoObjects/PseudoParticle.cpp
@@ -14,7 +14,7 @@ PseudoParticle::PseudoParticle() :
 
 }
 
-PseudoParticle::PseudoParticle(float phi, float eta, float pt, float theta) :
+PseudoParticle::PseudoParticle(double phi, double eta, double pt, double theta) :
     Phi(phi), Eta(eta), Pt(pt), Theta(theta), D0(9999), Charge(-5555) {
 
 }
@@ -22,79 +22,79 @@ PseudoParticle::PseudoParticle(float phi, float eta, float pt, float theta) :
 PseudoParticle::~PseudoParticle() {
 }
 
-void PseudoParticle::setPhi(float phi) {
+void PseudoParticle::setPhi(double phi) {
     Phi = phi;
 }
 
-void PseudoParticle::setEta(float eta) {
+void PseudoParticle::setEta(double eta) {
     Eta = eta;
 }
 
-void PseudoParticle::setPt(float pt) {
+void PseudoParticle::setPt(double pt) {
     Pt = pt;
 }
 
-void PseudoParticle::setTheta(float theta) {
+void PseudoParticle::setTheta(double theta) {
     Theta = theta;
 }
 
-void PseudoParticle::setD0(float d0) {
+void PseudoParticle::setD0(double d0) {
     D0 = d0;
 }
 
-void PseudoParticle::setCharge(float charge) {
+void PseudoParticle::setCharge(double charge) {
     Charge = charge;
 }
 
-float PseudoParticle::phi() const {
+double PseudoParticle::phi() const {
     return Phi;
 }
 
-float PseudoParticle::eta() const {
+double PseudoParticle::eta() const {
     return Eta;
 }
 
-float PseudoParticle::pt() const {
+double PseudoParticle::pt() const {
     return Pt;
 }
 
-float PseudoParticle::theta() const {
+double PseudoParticle::theta() const {
     return Theta;
 }
 
-float PseudoParticle::d0() const {
+double PseudoParticle::d0() const {
     return D0;
 }
 
-float PseudoParticle::charge() const {
+double PseudoParticle::charge() const {
     return Charge;
 }
 
-float PseudoParticle::deltaR(const PseudoParticlePointer otherParticle) const {
+double PseudoParticle::deltaR(const PseudoParticlePointer otherParticle) const {
     return deltaR(Phi, Eta, otherParticle->phi(), otherParticle->eta());
 }
 
-float PseudoParticle::deltaR(const ParticlePointer otherParticle) const {
+double PseudoParticle::deltaR(const ParticlePointer otherParticle) const {
     return deltaR(Phi, Eta, otherParticle->phi(), otherParticle->eta());
 }
 
-float PseudoParticle::deltaR(float phi1, float eta1, float phi2, float eta2) const {
+double PseudoParticle::deltaR(double phi1, double eta1, double phi2, double eta2) const {
     return sqrt(pow(deltaPhi(phi1, phi2), 2) + pow(deltaEta(eta1, eta2), 2));
 }
 
-float PseudoParticle::deltaPhi(float phi1, float phi2) const {
-    float delPhi(999.0);
+double PseudoParticle::deltaPhi(double phi1, double phi2) const {
+    double delPhi(999.0);
     delPhi = fabs(phi1 - phi2);
     if (delPhi > TMath::Pi())
         delPhi = 2.0 * TMath::Pi() - delPhi;
     return delPhi;
 }
 
-float PseudoParticle::deltaEta(float eta1, float eta2) const{
+double PseudoParticle::deltaEta(double eta1, double eta2) const{
     return eta1 - eta2;
 }
 
-bool PseudoParticle::isWithinDeltaR(float delta_R, const PseudoParticlePointer particle) const{
+bool PseudoParticle::isWithinDeltaR(double delta_R, const PseudoParticlePointer particle) const{
     return deltaR(particle) < delta_R;
 }
